v1.4.7 — address Bitfocus code review feedback
Structural (H2): split monolithic main.js into src/ modules
- src/main.js — core class (connection, lifecycle, message handler)
- src/actions.js — all action definitions
- src/feedbacks.js — recording_state and tally_state_active feedbacks
- src/variables.js — static variable definitions
- src/upgrades.js — upgrade script (v1.4.0 action ID rename + feedback cleanup)

Bug fixes: ws error handler now calls updateStatus(ConnectionFailure),
stagger timers tracked in Set and cleared on destroy, ISO/FPS NaN
guards, LUT subscribe params corrected to ENABLE_CAMERA_LUT_SDI_1/2

Config: add engines + packageManager fields, restore scripts/repository,
.gitattributes (LF), .prettierignore, .yarnrc.yml (nodeLinker),
LICENSE updated to Bitfocus attribution, .gitignore updated
