+++ .prettierignore
@@ -1,0 +1,2 @@
+package.json
+/LICENSE.md
