+++ companion/manifest.json
@@ -1,9 +1,10 @@
 {
+    "$schema": "../node_modules/@companion-module/base/assets/manifest.schema.json",
     "id": "red-rcp2",
     "name": "RED RCP2 Camera Control",
     "shortname": "RCP2-Control",
     "description": "RCP2 Red Camera Control via websocket",
-    "version": "1.4.6",
+    "version": "0.0.0",
     "license": "MIT",
     "repository": "git+https://github.com/bitfocus/companion-module-red-rcp2.git",
     "bugs": "https://github.com/bitfocus/companion-module-red-rcp2/issues",
@@ -18,7 +19,7 @@
         "type": "node22",
         "api": "nodejs-ipc",
         "apiVersion": "0.0.0",
-        "entrypoint": "../main.js"
+        "entrypoint": "../src/main.js"
     },
     "manufacturer": "RED",
     "products": [
@@ -26,10 +27,7 @@
     ],
     "keywords": [
         "websocket",
-        "RED",
         "Raptor",
-        "R3D",
-        "RCP",
-        "RCP2"
+        "R3D"
     ]
 }
